refactor(test): move BTC transactions mock
